+++ src/components/ApiInfo/index.ts
@@ -1,2 +1,1 @@
-export { ApiDescription } from './ApiDescription';
 export { ApiInfo } from './ApiInfo';