+++ src/components/ContentItems/ContentItems.tsx
@@ -1,47 +1,28 @@
 import { observer } from 'mobx-react';
 import * as React from 'react';
 
-import { SECTION_ATTR } from '../../services/MenuStore';
-import { Markdown } from '../Markdown/Markdown';
+import { AdvancedMarkdown } from '../Markdown/AdvancedMarkdown';
 
-import { H1, H2, MiddlePanel, Row, ShareLink } from '../../common-elements';
-import { MDXComponentMeta } from '../../services/MarkdownRenderer';
+import { H1, H2, MiddlePanel, Row, Section, ShareLink } from '../../common-elements';
 import { ContentItemModel } from '../../services/MenuBuilder';
 import { GroupModel, OperationModel } from '../../services/models';
 import { Operation } from '../Operation/Operation';
-import { SecurityDefs } from '../SecuritySchemes/SecuritySchemes';
-import { StoreConsumer } from '../StoreBuilder';
 
 @observer
 export class ContentItems extends React.Component<{
   items: ContentItemModel[];
-  allowedMdComponents?: Dict<MDXComponentMeta>;
 }> {
-  static defaultProps = {
-    allowedMdComponents: {
-      'security-definitions': {
-        component: SecurityDefs,
-        propsSelector: _store => ({
-          securitySchemes: _store!.spec.securitySchemes,
-        }),
-      },
-    },
-  };
-
   render() {
     const items = this.props.items;
     if (items.length === 0) {
       return null;
     }
-    return items.map(item => (
-      <ContentItem item={item} key={item.id} allowedMdComponents={this.props.allowedMdComponents} />
-    ));
+    return items.map(item => <ContentItem item={item} key={item.id} />);
   }
 }
 
 export interface ContentItemProps {
   item: ContentItemModel;
-  allowedMdComponents?: Dict<MDXComponentMeta>;
 }
 
 @observer
@@ -67,39 +48,37 @@ export class ContentItem extends React.Component<ContentItemProps> {
         throw new Error('Unknown item type');
     }
 
-    return [
-      <div key="section" {...{ [SECTION_ATTR]: item.id }}>
-        {content}
-      </div>,
-      (item as any).items && <ContentItems key="content" items={(item as any).items} />,
-    ];
+    return (
+      <>
+        <Section id={item.id} underlined={item.type === 'operation'}>
+          {content}
+        </Section>
+        {item.items && <ContentItems items={item.items} />}
+      </>
+    );
   }
 }
+
+const middlePanelWrap = component => <MiddlePanel>{component}</MiddlePanel>;
 
 @observer
 export class SectionItem extends React.Component<ContentItemProps> {
   render() {
     const { name, description, level } = this.props.item as GroupModel;
-    const components = this.props.allowedMdComponents;
+
     const Header = level === 2 ? H2 : H1;
     return (
-      <Row>
-        <MiddlePanel>
-          <Header>
-            <ShareLink href={'#' + this.props.item.id} />
-            {name}
-          </Header>
-          {components ? (
-            <StoreConsumer>
-              {store => (
-                <Markdown source={description || ''} allowedComponents={components} store={store} />
-              )}
-            </StoreConsumer>
-          ) : (
-            <Markdown source={description || ''} />
-          )}
-        </MiddlePanel>
-      </Row>
+      <>
+        <Row>
+          <MiddlePanel>
+            <Header>
+              <ShareLink to={this.props.item.id} />
+              {name}
+            </Header>
+          </MiddlePanel>
+        </Row>
+        <AdvancedMarkdown source={description || ''} htmlWrap={middlePanelWrap} />
+      </>
     );
   }
 }