+++ src/components/Fields/FieldDetails.tsx
@@ -28,7 +28,14 @@ export class FieldDetails extends React.PureComponent<FieldProps> {
         <div>
           <TypePrefix>{schema.typePrefix}</TypePrefix>
           <TypeName>{schema.displayType}</TypeName>
-          {schema.displayFormat && <TypeFormat> &lt;{schema.displayFormat}&gt; </TypeFormat>}
+          {schema.displayFormat && (
+            <TypeFormat>
+              {' '}
+              &lt;
+              {schema.displayFormat}
+              &gt;{' '}
+            </TypeFormat>
+          )}
           {schema.title && <TypeTitle> ({schema.title}) </TypeTitle>}
           <ConstraintsView constraints={schema.constraints} />
           {schema.nullable && <NullableLabel> Nullable </NullableLabel>}