+++ src/services/MarkdownRenderer.ts
@@ -2,6 +2,7 @@ import * as marked from 'marked';
 
 import { highlight, safeSlugify } from '../utils';
 import { AppStore } from './AppStore';
+import { RedocNormalizedOptions } from './RedocNormalizedOptions';
 
 const renderer = new marked.Renderer();
 
@@ -12,7 +13,7 @@ marked.setOptions({
   },
 });
 
-export const LEGACY_REGEXP = '^\\s*<!-- ReDoc-Inject:\\s+?{component}\\s+?-->\\s*$';
+export const LEGACY_REGEXP = '^\\s*<!-- ReDoc-Inject:\\s+?<{component}\\s*?/?>\\s+?-->\\s*$';
 export const MDX_COMPONENT_REGEXP = '^\\s*<{component}\\s*?/>\\s*$';
 export const COMPONENT_REGEXP = '(?:' + LEGACY_REGEXP + '|' + MDX_COMPONENT_REGEXP + ')';
 
@@ -35,13 +36,21 @@ export function buildComponentComment(name: string) {
 }
 
 export class MarkdownRenderer {
+  static containsComponent(rawText: string, componentName: string) {
+    const anyCompRegexp = new RegExp(
+      COMPONENT_REGEXP.replace(/{component}/g, componentName),
+      'gmi',
+    );
+    return anyCompRegexp.test(rawText);
+  }
+
   headings: MarkdownHeading[] = [];
   currentTopHeading: MarkdownHeading;
 
   private headingEnhanceRenderer: marked.Renderer;
   private originalHeadingRule: typeof marked.Renderer.prototype.heading;
 
-  constructor() {
+  constructor(public options?: RedocNormalizedOptions) {
     this.headingEnhanceRenderer = new marked.Renderer();
     this.originalHeadingRule = this.headingEnhanceRenderer.heading.bind(
       this.headingEnhanceRenderer,
@@ -140,15 +149,17 @@ export class MarkdownRenderer {
 
   // TODO: rewrite this completelly! Regexp-based 👎
   // Use marked ecosystem
-  renderMdWithComponents(
-    rawText: string,
-    components: Dict<MDXComponentMeta>,
-  ): Array<string | MDXComponentMeta> {
+  renderMdWithComponents(rawText: string): Array<string | MDXComponentMeta> {
+    const components = this.options && this.options.allowedMdComponents;
+    if (!components || Object.keys(components).length === 0) {
+      return [this.renderMd(rawText)];
+    }
+
     const componentDefs: string[] = [];
     const names = '(?:' + Object.keys(components).join('|') + ')';
 
     const anyCompRegexp = new RegExp(
-      COMPONENT_REGEXP.replace(/{component}/g, '(<?' + names + '.*?)'),
+      COMPONENT_REGEXP.replace(/{component}/g, '(' + names + '.*?)'),
       'gmi',
     );
     let match = anyCompRegexp.exec(rawText);
@@ -189,10 +200,6 @@ function parseComponent(
   componentName?: string;
   attrs: any;
 } {
-  if (htmlTag.startsWith('<')) {
-    return legacyParse(htmlTag);
-  }
-
   const match = /([\w_-]+)(\s+[\w_-]+\s*={[^}]*?})*/.exec(htmlTag);
   if (match === null || match.length <= 1) {
     return { componentName: undefined, attrs: {} };
@@ -214,22 +221,5 @@ function parseComponent(
   return {
     componentName,
     attrs,
-  };
-}
-
-function legacyParse(
-  htmlTag: string,
-): {
-  componentName?: string;
-  attrs: any;
-} {
-  const match = /<([\w_-]+).*?>/.exec(htmlTag);
-  if (match === null || match.length <= 1) {
-    return { componentName: undefined, attrs: {} };
-  }
-  const componentName = match[1];
-  return {
-    componentName,
-    attrs: {}, // TODO
   };
 }